+++ src/main/java/roomescape/controller/ReservationController.java
@@ -10,10 +10,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.CreateReservationRequest;
-import roomescape.dto.ReservationCreation;
-import roomescape.dto.ReservationResponse;
+import roomescape.controller.dto.request.CreateReservationRequest;
+import roomescape.controller.dto.response.ReservationResponse;
 import roomescape.service.ReservationService;
+import roomescape.service.dto.ReservationCreation;
 
 @RestController
 @RequestMapping("/reservations")
