+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -11,10 +11,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.CreateReservationTimeRequest;
-import roomescape.dto.ReservationTimeCreation;
-import roomescape.dto.ReservationTimeResponse;
+import roomescape.controller.dto.request.CreateReservationTimeRequest;
+import roomescape.controller.dto.response.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
+import roomescape.service.dto.ReservationTimeCreation;
 
 @RestController
 @RequestMapping("/times")
