+++ src/main/java/roomescape/controller/dto/request/CreateReservationTimeRequest.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto.request;
 
 import java.time.LocalTime;
