+++ src/main/java/roomescape/controller/dto/response/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.controller.dto.response;
 
 import java.time.LocalTime;
 import roomescape.domain.reservationtime.ReservationTime;
